Handle cronjob logic in role
This is semi complex dict layout which makes it more extensible down the road.
The trade off seems worth it to me but that is up for debate. I'm also sure
there is cleaner way to store and loop over the cronjob environment vars but I
was hitting walls and decided to go with good ol' lists.

Take note that the environments vars were added because of issues encountered
with python3, ascii encoding, and the click library being used by curator. See
http://click.pocoo.org/5/python3/#python-3-surrogate-handling for details.